risk_management/base_risk_manager.py
from typing import Dict, Any, Optional
import numpy as np
import pandas as pd
from riskfolio import RiskFunctions

class BaseRiskManager:
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.risk_functions = RiskFunctions()
        self.portfolio = {}
        self.risk_metrics = {}
        self._initialize_risk_parameters()

    def _initialize_risk_parameters(self):
        """
        Initialize risk parameters from config
        """
        self.max_drawdown = self.config.get('max_drawdown', 0.2)  # 20%
        self.position_size_limit = self.config.get('position_size_limit', 0.1)  # 10%
        self.volatility_target = self.config.get('volatility_target', 0.15)  # 15%
        self.risk_free_rate = self.config.get('risk_free_rate', 0.02)  # 2%

    def calculate_position_size(self, symbol: str, price: float, 
                              portfolio_value: float) -> float:
        """
        Calculate position size based on risk parameters
        """
        max_position_value = portfolio_value * self.position_size_limit
        return min(max_position_value / price, 
                  self._calculate_volatility_adjusted_size(symbol))

    def _calculate_volatility_adjusted_size(self, symbol: str) -> float:
        """
        Calculate position size adjusted for volatility
        """
        # Placeholder implementation
        return 100  # Default fixed size

    def check_risk_limits(self, portfolio: Dict[str, Any]) -> bool:
        """
        Check if portfolio is within risk limits
        """
        self._update_portfolio_metrics(portfolio)
        
        if self.risk_metrics['drawdown'] > self.max_drawdown:
            return False
        
        if self.risk_metrics['volatility'] > self.volatility_target:
            return False
            
        return True

    def _update_portfolio_metrics(self, portfolio: Dict[str, Any]):
        """
        Update portfolio risk metrics
        """
        returns = self._calculate_portfolio_returns(portfolio)
        self.risk_metrics['drawdown'] = self.risk_functions.max_drawdown(returns)
        self.risk_metrics['volatility'] = np.std(returns)
        self.risk_metrics['sharpe_ratio'] = self.risk_functions.sharpe_ratio(
            returns, self.risk_free_rate)

    def _calculate_portfolio_returns(self, portfolio: Dict[str, Any]) -> pd.Series:
        """
        Calculate portfolio returns
        """
        # Placeholder implementation
        return pd.Series([0.01, -0.02, 0.03])  # Example returns

    def get_risk_metrics(self) -> Dict[str, float]:
        """
        Get current risk metrics
        """
        return self.risk_metrics

    def apply_risk_controls(self, signals: Dict[str, str], 
                          portfolio: Dict[str, Any]) -> Dict[str, str]:
        """
        Apply risk controls to trading signals
        """
        if not self.check_risk_limits(portfolio):
            # If risk limits are exceeded, close all positions
            return {symbol: 'sell' for symbol in portfolio['positions'].keys()}
        
        return signals
