requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Data processing
yfinance>=0.1.70
requests>=2.26.0
websockets>=10.0

# Backtesting
backtrader>=1.9.76.123

# Risk management
riskfolio-lib>=3.3.0

# Portfolio optimization
skfolio>=0.1.0

# Optimization
scikit-optimize>=0.9.0

# Visualization
matplotlib>=3.4.0
plotly>=5.3.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Development
black>=21.7b0
flake8>=3.9.0
mypy>=0.910

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# Logging
loguru>=0.5.0

# Configuration
pyyaml>=5.4.0
